Correction bug request avec auth

+++ controllers/books.js
@@ -1,5 +1,6 @@
-const { request } = require('../app');
+//const { request } = require('../app');
 const Book = require('../models/Book');
+const fs = require('fs');
 
 exports.createBook = (req, res, next) => {
     const bookObject = JSON.parse(req.body.book);
@@ -37,9 +38,22 @@ exports.modifyBook = (req, res, next) => {
 };
 
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Book deleted'}))
-        .catch(error => res.status(400).json({ error }));
+    Book.findOne({ _id: req.params.id})
+        .then(book => {
+            if(book.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Unauthorized'})
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({ _id: req.params.id})
+                        .then(() => {res.status(200).json({ message: 'Book deleted'})})
+                        .catch(error => res.status(401).json({ error }));
+                })
+            };
+        })
+        .catch(error => {
+            res.status(500).json({ error })
+        });
 };
 
 exports.getOneBook = (req, res, next) => {
